test (#19)
* fix: modules

* move: damage_resistance_capability

* move: egg_capability

* move: size_capability

* move: world_tracker_capability

* move: skill_capability

* move: food_capability

* ref: CTItemRegistry

* move: alloy_calculator

* fix: Settings

* ref: all

* build(deps): bump actions/setup-java in the dependencies group

Bumps the dependencies group with 1 update: [actions/setup-java](https://github.com/actions/setup-java).


Updates `actions/setup-java` from 4.6.0 to 4.7.0
- [Release notes](https://github.com/actions/setup-java/releases)
- [Commits](https://github.com/actions/setup-java/compare/v4.6.0...v4.7.0)

---
updated-dependencies:
- dependency-name: actions/setup-java
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: dependencies
...


* move: animal

* move: animal items

* fix: playerdata

* move: metal_capability

* move: forge_capability

* revert: move animal items

* move: flask

* fix: misc

* add: textures mc

* move: all

* move: bear_trap & snare

* move: bellows

* move: blast_furnace & bloomery & molten

* move: bloom

* move: charcoal_forge

* move: crucible

* move: charcoal_pile

* move: firepit & wood_ash

* move: thatch & straw & pit_kiln

* move: ItemTooltip

* move: ItemTooltip 2

* build(deps): bump actions/upload-artifact in the dependencies group

Bumps the dependencies group with 1 update: [actions/upload-artifact](https://github.com/actions/upload-artifact).


Updates `actions/upload-artifact` from 4.6.0 to 4.6.1
- [Release notes](https://github.com/actions/upload-artifact/releases)
- [Commits](https://github.com/actions/upload-artifact/compare/v4.6.0...v4.6.1)

---
updated-dependencies:
- dependency-name: actions/upload-artifact
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: dependencies
...


* ref: base api

* move: smeltery_firebox & smeltery_cauldron

* move: latex_extractor & infected_air & freeze_dryer & shard

* move: ice_bunker

* fix: FreezeDryer

* fix: armor textures

* add: top provider

* fix: NetworkThreadedWrapper

* ref: button

* fix: ItemFireStarter

* move: TFG-Resourcepacks-1.12.2 pt1

* move: TFG-Resourcepacks-1.12.2 pt2

* move: TFG-Resourcepacks-1.12.2 pt3

* move: TFG-Resourcepacks-1.12.2 pt4

* move: other assets modpack

* move: powderkeg

* move: recipe

* move: recipe pt2

* ref: gt material

* move: helper

* fix: api

* move: hotornot

* build(deps): bump actions/upload-artifact in the dependencies group

Bumps the dependencies group with 1 update: [actions/upload-artifact](https://github.com/actions/upload-artifact).


Updates `actions/upload-artifact` from 4.6.1 to 4.6.2
- [Release notes](https://github.com/actions/upload-artifact/releases)
- [Commits](https://github.com/actions/upload-artifact/compare/v4.6.1...v4.6.2)

---
updated-dependencies:
- dependency-name: actions/upload-artifact
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: dependencies
...


* build(deps): bump actions/setup-java in the dependencies group

Bumps the dependencies group with 1 update: [actions/setup-java](https://github.com/actions/setup-java).


Updates `actions/setup-java` from 4.7.0 to 4.7.1
- [Release notes](https://github.com/actions/setup-java/releases)
- [Commits](https://github.com/actions/setup-java/compare/v4.7.0...v4.7.1)

---
updated-dependencies:
- dependency-name: actions/setup-java
  dependency-version: 4.7.1
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: dependencies
...


* update gradle

* начало миграции

регистрация на Supplier<V>

* миграция pt2

* миграция pt3

* миграция pt4

* миграция pt5

* ref: network

* ref: feature(Size, Egg, AdvancedData)

* ref: feature(Ambiental)

---------

+++ src/main/java/net/dries007/tfc/objects/blocks/agriculture/BlockCropDead.java
@@ -1,9 +1,8 @@
-/*
- * Work under Copyright. Licensed under the EUPL.
- * See the project README.md and LICENSE.txt for more information.
- */
-
 package net.dries007.tfc.objects.blocks.agriculture;
+
+import su.terrafirmagreg.modules.core.capabilities.playerdata.CapabilityPlayerData;
+import su.terrafirmagreg.modules.core.feature.skill.SimpleSkill;
+import su.terrafirmagreg.modules.core.feature.skill.SkillType;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockBush;
@@ -24,17 +23,14 @@ import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.common.EnumPlantType;
 
-import net.dries007.tfc.api.capability.player.CapabilityPlayerData;
 import net.dries007.tfc.api.types.ICrop;
 import net.dries007.tfc.objects.blocks.BlocksTFC;
 import net.dries007.tfc.objects.blocks.plants.BlockEmergentTallWaterPlantTFC;
 import net.dries007.tfc.objects.blocks.plants.BlockWaterPlantTFC;
 import net.dries007.tfc.objects.items.ItemSeedsTFC;
 import net.dries007.tfc.util.agriculture.Crop;
-import net.dries007.tfc.util.skills.SimpleSkill;
-import net.dries007.tfc.util.skills.SkillType;
 import net.dries007.tfc.world.classic.ChunkGenTFC;
-import tfcflorae.objects.blocks.plants.BlockWaterPlantTFCF;
+import net.dries007.tfc.objects.blocks.plants.BlockWaterPlantTFCF;
 
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
@@ -152,7 +148,7 @@ public class BlockCropDead extends BlockBush { //implements IGrowingPlant
           IBlockState stateDown = worldIn.getBlockState(pos.down());
           Material material = stateDown.getMaterial();
           return soil.getBlock().canSustainPlant(soil, worldIn, pos.down(), EnumFacing.UP, this)
-                 || material == Material.WATER && (Integer) stateDown.getValue(BlockLiquid.LEVEL) == 0 && stateDown == ChunkGenTFC.FRESH_WATER
+                 || material == Material.WATER && stateDown.getValue(BlockLiquid.LEVEL) == 0 && stateDown == ChunkGenTFC.FRESH_WATER
                  || material == Material.ICE || material == Material.CORAL && !(state.getBlock() instanceof BlockEmergentTallWaterPlantTFC);
         }
       } else {
